Update README.md
Percy Jackson: Sea of Monsters (also known as Percy Jackson & the Olympians: The Sea of Monsters and Percy Jackson and the Sea of Monsters) is a 2013 fantasy adventure film directed by Thor Freudenthal from a screenplay by Marc Guggenheim, based on the 2006 novel The Sea of Monsters by Rick Riordan. The sequel to Percy Jackson & the Olympians: The Lightning Thief (2010), it is the second and final installment in the Percy Jackson film series.